Extract normaliseErrorListTriggers test helper
